+++ src/components/Dashboard/Dashboard.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { Redirect } from 'react-router-dom';
 
 import { Auth } from '../../fixtures/types';
 import { TaskList, UserPanel } from '../index';
