+++ src/components/Task/index.ts
@@ -1,0 +1,7 @@
+import TaskCreate from './TaskCreate/TaskCreate';
+import TaskEdit from './TaskEdit/TaskEdit';
+import TaskItem from './TaskItem/TaskItem';
+import TaskList from './TaskList/TaskList';
+import TaskRemove from './TaskRemove/TaskRemove';
+
+export { TaskCreate, TaskEdit, TaskItem, TaskList, TaskRemove };
