+++ src/components/index.ts
@@ -5,11 +5,7 @@ import RandomAvatar from './RandomAvatar/RandomAvatar';
 import SearchBar from './SearchBar/SearchBar';
 import SignInComponent from './SignInComponent/SignInComponent';
 import SignUpComponent from './SignUpComponent/SignUpComponent';
-import TaskCreate from './TaskCreate/TaskCreate';
-import TaskEdit from './TaskEdit/TaskEdit';
-import TaskItem from './TaskItem/TaskItem';
-import TaskList from './TaskList/TaskList';
-import TaskRemove from './TaskRemove/TaskRemove';
+import { TaskCreate, TaskEdit, TaskItem, TaskList, TaskRemove } from './Task';
 import UserPanel from './UserPanel/UserPanel';
 
 export {
